Finished middleware-v2, extracted.
